Fixed problem caused by order of emitters calls.
If some `sync_se_emitter`s were called after `sync_test_emitter`s, the
corresponding targets are not blocked from running at the same time as
the synchronous tasks.
The solution is to store a list of targets for which `sync_test_emitter`
was already called and add additional `SideEffect`s for them whenever
`sync_se_emitter` is called for a new target.